# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Go Tasks API
  description: |
    Task and habit management API. Users register and manage their own tasks,
    categories, daily logs, and answers. All data is scoped to the authenticated user.

    ## Authentication

    Protected endpoints require a valid JWT access token in the `Authorization` header
    using the Bearer scheme. Access tokens are RS256-signed and valid for 15 minutes.

    **Token Claims:**
    - `sub` — user ID (UUID)
    - `iss` — issuer (configured via JWT_ISSUER)
    - `aud` — audience (configured via JWT_AUDIENCE)
    - `exp` — expiration time
    - `nbf` — not before time
    - `iat` — issued at time
    - `jti` — unique token ID (UUID)

    ## Refresh Tokens

    Refresh tokens are valid for 1 hour and support rotation. On use, the old token
    is deleted and a new one is issued. Logout revokes the refresh token and adds
    the access token's jti to a blocklist.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/tasks:
    get:
      tags:
        - Task
      summary: List tasks
      operationId: listTasks
      parameters:
        - name: category_id
          in: query
          description: Filter by category ID
          schema:
            type: string
            format: uuid
        - name: active
          in: query
          description: Filter by active status (default true)
          schema:
            type: boolean
            default: true
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Task
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithDetails"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Category not found
  /api/v1/tasks/{id}:
    get:
      tags:
        - Task
      summary: Get a task with details
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Task with schedule and options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithDetails"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Task
      summary: Update a task
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Task
      summary: Soft delete a task
      description: Sets is_active to false. Returns 409 if already inactive.
      operationId: deleteTask
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Task deactivated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Task is already inactive
  /api/v1/tasks/{id}/permanent:
    delete:
      tags:
        - Task
      summary: Permanently delete a task
      description: Hard deletes an inactive task and all its related data. Returns 409 if still active.
      operationId: permanentDeleteTask
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Task permanently deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot permanently delete an active task
  /api/v1/tasks/{id}/reactivate:
    post:
      tags:
        - Task
      summary: Reactivate a task
      description: Sets is_active to true. Returns 409 if already active.
      operationId: reactivateTask
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Task reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Task is already active
  /api/v1/tasks/bulk-delete:
    post:
      tags:
        - Task
      summary: Bulk soft delete tasks
      description: Soft deletes multiple tasks. Inactive IDs are ignored.
      operationId: bulkDeleteTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          description: Invalid request (empty list, too many IDs, invalid UUID)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/tasks/bulk-permanent-delete:
    post:
      tags:
        - Task
      summary: Bulk permanently delete tasks
      description: Hard deletes multiple inactive tasks. Active IDs are ignored.
      operationId: bulkPermanentDeleteTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk permanent delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkPermanentDeleteResponse"
        "400":
          description: Invalid request (empty list, too many IDs, invalid UUID)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/tasks/inactive:
    get:
      tags:
        - Task
      summary: List inactive tasks
      description: Lists all soft-deleted tasks regardless of category status.
      operationId: listInactiveTasks
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of inactive tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Occurrence endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT access token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        answer_type:
          type: string
          enum: [string, integer, boolean, select]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        recurrence_type:
          type: string
          enum: [once, daily, every_n_days, weekly, every_n_weeks, monthly_date, monthly_weekday, yearly]
        recurrence_interval:
          type: integer
        days_of_week:
          type: array
          items:
            type: integer
        month_day:
          type: integer
        month_week:
          type: integer
        month_weekday:
          type: integer
        month_of_year:
          type: integer
        scheduled_times:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date
        end_type:
          type: string
          enum: [never, on_date, after_n]
        end_date:
          type: string
          format: date
        end_after_n:
          type: integer
        created_at:
          type: string
          format: date-time
    TaskSelectOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        value:
          type: string
          maxLength: 100
        position:
          type: integer
        created_at:
          type: string
          format: date-time
    TaskWithDetails:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/Task"
        schedule:
          $ref: "#/components/schemas/TaskSchedule"
        select_options:
          type: array
          items:
            $ref: "#/components/schemas/TaskSelectOption"
    CreateTaskRequest:
      type: object
      required:
        - category_id
        - name
        - answer_type
        - schedule
      properties:
        category_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        answer_type:
          type: string
          enum: [string, integer, boolean, select]
        schedule:
          $ref: "#/components/schemas/ScheduleRequest"
        select_options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            type: object
            properties:
              value:
                type: string
                maxLength: 100
    ScheduleRequest:
      type: object
      required:
        - recurrence_type
        - start_date
      properties:
        recurrence_type:
          type: string
          enum: [once, daily, every_n_days, weekly, every_n_weeks, monthly_date, monthly_weekday, yearly]
        recurrence_interval:
          type: integer
          minimum: 1
        days_of_week:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
        month_day:
          type: integer
          minimum: 1
          maximum: 31
        month_week:
          type: integer
          minimum: 1
          maximum: 5
        month_weekday:
          type: integer
          minimum: 0
          maximum: 6
        month_of_year:
          type: integer
          minimum: 1
          maximum: 12
        scheduled_times:
          type: array
          items:
            type: string
            pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        start_date:
          type: string
          format: date
        end_type:
          type: string
          enum: [never, on_date, after_n]
        end_date:
          type: string
          format: date
        end_after_n:
          type: integer
          minimum: 1
    UpdateTaskRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000

    # Occurrence schemas
...
    BulkDeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
          description: List of IDs to delete (1-100 IDs, duplicates deduplicated server-side)
    BulkDeleteResponse:
      type: object
      properties:
        requested:
          type: integer
          description: Number of unique IDs requested (after deduplication)
        soft_deleted:
          type: integer
          description: Number of items soft deleted (set to inactive)
    BulkPermanentDeleteResponse:
      type: object
      properties:
        requested:
          type: integer
          description: Number of unique IDs requested (after deduplication)
        permanently_deleted:
          type: integer
          description: Number of items permanently deleted
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Resource not found